+++ user-service/src/main/java/com/financialmonitoring/userservice/adapter/factory/UserFactoryImpl.java
@@ -3,7 +3,6 @@ package com.financialmonitoring.userservice.adapter.factory;
 import com.financialmonitoring.userservice.adapter.dto.RegisterRequestDTO;
 import com.financialmonitoring.userservice.infra.model.Role;
 import com.financialmonitoring.userservice.infra.model.User;
-import com.financialmonitoring.userservice.domain.port.factory.UserFactory;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
