+++ user-service/src/main/java/com/financialmonitoring/userservice/domain/service/AuthService.java
@@ -4,13 +4,13 @@ import com.financialmonitoring.userservice.adapter.dto.*;
 import com.financialmonitoring.userservice.infra.model.Role;
 import com.financialmonitoring.userservice.infra.model.User;
 import com.financialmonitoring.userservice.config.exception.BadRequestException;
-import com.financialmonitoring.userservice.domain.port.factory.UserFactory;
-import com.financialmonitoring.userservice.domain.port.in.GetUserByEmailUseCase;
-import com.financialmonitoring.userservice.domain.port.in.LoginUseCase;
-import com.financialmonitoring.userservice.domain.port.in.RegisterUserUseCase;
-import com.financialmonitoring.userservice.domain.port.in.VerifyTokenUseCase;
-import com.financialmonitoring.userservice.domain.port.out.AuthRepositoryPort;
-import com.financialmonitoring.userservice.domain.port.out.RoleRepositoryPort;
+import com.financialmonitoring.userservice.adapter.factory.UserFactory;
+import com.financialmonitoring.userservice.port.input.GetUserByEmailUseCase;
+import com.financialmonitoring.userservice.port.input.LoginUseCase;
+import com.financialmonitoring.userservice.port.input.RegisterUserUseCase;
+import com.financialmonitoring.userservice.port.input.VerifyTokenUseCase;
+import com.financialmonitoring.userservice.port.output.AuthRepositoryPort;
+import com.financialmonitoring.userservice.port.output.RoleRepositoryPort;
 import com.financialmonitoring.userservice.domain.utils.TokenUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
