+++ user-service/src/main/java/com/financialmonitoring/userservice/port/input/VerifyTokenUseCase.java
@@ -1,4 +1,4 @@
-package com.financialmonitoring.userservice.domain.port.in;
+package com.financialmonitoring.userservice.port.input;
 
 public interface VerifyTokenUseCase {
     boolean verifyToken(String authHeader);
