+++ user-service/src/test/java/com/financialmonitoring/userservice/domain/service/AuthServiceTest.java
@@ -7,9 +7,9 @@ import com.financialmonitoring.userservice.adapter.dto.RegisterResponseDTO;
 import com.financialmonitoring.userservice.infra.model.Role;
 import com.financialmonitoring.userservice.infra.model.User;
 import com.financialmonitoring.userservice.config.exception.BadRequestException;
-import com.financialmonitoring.userservice.domain.port.factory.UserFactory;
-import com.financialmonitoring.userservice.domain.port.out.AuthRepositoryPort;
-import com.financialmonitoring.userservice.domain.port.out.RoleRepositoryPort;
+import com.financialmonitoring.userservice.adapter.factory.UserFactory;
+import com.financialmonitoring.userservice.port.output.AuthRepositoryPort;
+import com.financialmonitoring.userservice.port.output.RoleRepositoryPort;
 import com.financialmonitoring.userservice.domain.utils.TokenUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
