+++ user-service/src/test/java/com/financialmonitoring/userservice/domain/service/UserServiceTest.java
@@ -2,7 +2,7 @@ package com.financialmonitoring.userservice.domain.service;
 
 import com.financialmonitoring.userservice.infra.model.Role;
 import com.financialmonitoring.userservice.infra.model.User;
-import com.financialmonitoring.userservice.domain.port.out.AuthRepositoryPort;
+import com.financialmonitoring.userservice.port.output.AuthRepositoryPort;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
